qa_fixtures/tests/requirements.txt
requests
pytest
retrying
psycopg2
selenium
pyyaml

